connectionstring/requirements.txt
azure-servicebus
requests